requirements.txt
boto3
flask
flask-cors
aws-wsgi
psycopg2-binary